fix: 修复时间类型

+++ src/work/externalContact/customerAcquisition/response/responseCreateLink.go
@@ -7,7 +7,7 @@ type Link struct {
 	LinkName   string `json:"link_name"` // 链接名称
 	Url        string `json:"url"`
 	SkipVerify bool   `json:"skip_verify"`
-	CreateTime int    `json:"create_time"`
+	CreateTime int64  `json:"create_time"`
 }
 
 type ResponseCreateLink struct {
